Reorganizando arquivos resumo do dia
