--- src/sections/Today/index.js
@@ -1,4 +1,0 @@
-import Tutela from './Tutela';
-import Pip from './Pip';
-
-export default { Tutela, Pip };
