+++ src/sections/index.js
@@ -2,5 +2,5 @@ export { default as Alerts } from './Alerts';
 export { default as PerformanceRadar } from './PerformanceRadar';
 export { default as ProcessingTime } from './ProcessingTime';
 export { default as ProcessList } from './ProcessList';
-export { default as Today } from './Today/Tutela';
+export { default as Today } from './Today';
 export { default as YourDesk } from './YourDesk';
